bd funcional
