+++ api/src/models/User.js
@@ -1,0 +1,12 @@
+const { DataTypes } = require('sequelize');
+// Exportamos una funcion que define el modelo
+// Luego le injectamos la conexion a sequelize.
+module.exports = (sequelize) => {
+  // defino el modelo
+  sequelize.define('User', {
+    
+  },{
+    timestamps: false
+  }
+  );
+};
