+++ api/src/routes/index.js
@@ -1,0 +1,15 @@
+const { Router } = require('express');
+// Importar todos los routers;
+// Ejemplo: const authRouter = require('./auth.js');
+// const modelWorker = require('./Worker.js')
+// const modelUser = require('./User.js')
+//Seguir agregando lo que falte....
+
+const router = Router();
+
+// Configurar los routers
+// Ejemplo: router.use('/auth', authRouter);
+// router.use('/worker',modelWorker)
+// router.use('/user',modelUser)
+
+module.exports = router;
